refactored folder structure of auth module and renamed profile to user, added uuid generation to creation of user
